+++ src/Classes/Game/CGameCtnBlock.ts
