+++ src/Classes/Game/CGameCtnBlockSkin.ts
@@ -1,4 +1,4 @@
-import Node from './Node';
+import Node from '../Node';
 
 /**
  * Skin of a block.
