+++ src/Classes/Game/CGameCtnChallengeParameters.ts
@@ -1,4 +1,4 @@
-import Node from './Node';
+import Node from '../Node';
 import CGameCtnGhost from './CGameCtnGhost';
 
 /**
