+++ src/index.ts
@@ -1,14 +1,14 @@
-import CGameCtnBlockSkin from './Classes/CGameCtnBlockSkin';
-import CGameCtnChallenge from './Classes/CGameCtnChallenge';
-import CGameCtnChallengeParameters from './Classes/CGameCtnChallengeParameters';
-import CGameCtnCollectorList from './Classes/CGameCtnCollectorList';
-import CGameCtnGhost from './Classes/CGameCtnGhost';
-import CGameCtnMediaClip from './Classes/CGameCtnMediaClip';
-import CGameCtnMediaTrack from './Classes/CGameCtnMediaTrack';
-import CGameCtnReplayRecord from './Classes/CGameCtnReplayRecord';
-import CGameGhost from './Classes/CGameGhost';
-import CGameWaypointSpecialProperty from './Classes/CGameWaypointSpecialProperty';
-import CPlugEntRecordData from './Classes/CPlugEntRecordData';
+import CGameCtnBlockSkin from './Classes/Game/CGameCtnBlockSkin';
+import CGameCtnChallenge from './Classes/Game/CGameCtnChallenge';
+import CGameCtnChallengeParameters from './Classes/Game/CGameCtnChallengeParameters';
+import CGameCtnCollectorList from './Classes/Game/CGameCtnCollectorList';
+import CGameCtnGhost from './Classes/Game/CGameCtnGhost';
+import CGameCtnMediaClip from './Classes/Game/CGameCtnMediaClip';
+import CGameCtnMediaTrack from './Classes/Game/CGameCtnMediaTrack';
+import CGameCtnReplayRecord from './Classes/Game/CGameCtnReplayRecord';
+import CGameGhost from './Classes/Game/CGameGhost';
+import CGameWaypointSpecialProperty from './Classes/Game/CGameWaypointSpecialProperty';
+import CPlugEntRecordData from './Classes/Plug/CPlugEntRecordData';
 
 import GBX from './GBX';
